+++ Packages/com.datea83.choyutilities/Runtime/Utils/UiCollection.cs
@@ -1,8 +1,12 @@
+using System;
 using Unity.Mathematics;
 using UnityEngine;
+using UnityEngine.UIElements;
 
 namespace EugeneC.Utilities {
+	
     public static partial class HelperCollection {
+	    
         public static void ActivateDisplay() {
             for (var i = 1; i < Display.displays.Length; i++) {
                 Display.displays[i].Activate();
@@ -31,5 +35,42 @@ namespace EugeneC.Utilities {
 
             float2 Rotate(float2 rel) => new(cos * rel.x - sin * rel.y, sin * rel.x + cos * rel.y);
         }
+
+        public static VisualElement GetVisualElement(this EVisualElements type)
+		{
+			return type switch
+			{
+				EVisualElements.Button => new Button(),
+				EVisualElements.DoubleField => new DoubleField(),
+				EVisualElements.DropdownField => new DropdownField(),
+				EVisualElements.EnumField => new EnumField(),
+				EVisualElements.FloatField => new FloatField(),
+				EVisualElements.Foldout => new Foldout(),
+				EVisualElements.IntegerField => new IntegerField(),
+				EVisualElements.Label => new Label(),
+				EVisualElements.ListView => new ListView(),
+				EVisualElements.LongField => new LongField(),
+				EVisualElements.MinMaxSlider => new MinMaxSlider(),
+				EVisualElements.MultiColumnListView => new MultiColumnListView(),
+				EVisualElements.MultiColumnTreeView => new MultiColumnTreeView(),
+				EVisualElements.ProgressBar => new ProgressBar(),
+				EVisualElements.RadioButton => new RadioButton(),
+				EVisualElements.RadioButtonGroup => new RadioButtonGroup(),
+				EVisualElements.RepeatButton => new RepeatButton(),
+				EVisualElements.Slider => new Slider(),
+				EVisualElements.SliderInt => new SliderInt(),
+				EVisualElements.TemplateContainer => new TemplateContainer(),
+				EVisualElements.TextElement => new TextElement(),
+				EVisualElements.Toggle => new Toggle(),
+				EVisualElements.TreeView => new TreeView(),
+				EVisualElements.UnsignedLongField => new UnsignedLongField(),
+				EVisualElements.Vector2Field => new Vector2Field(),
+				EVisualElements.Vector2IntField => new Vector2IntField(),
+				EVisualElements.Vector3Field => new Vector3Field(),
+				EVisualElements.Vector3IntField => new Vector3IntField(),
+				EVisualElements.Vector4Field => new Vector4Field(),
+				_ => throw new ArgumentOutOfRangeException(nameof(type), type, "This enum value is editor-only or unsupported at runtime.")
+			};
+		}
     }
 }